Clean up solution folders, names, and namespaces
