+++ ICL.Core/AgentSystem/ICLBeamAgentSystem.cs
@@ -5,7 +5,7 @@ using Grasshopper.Kernel.Geometry.Voronoi;
 using ABxM.Core.Agent;
 using ABxM.Core.AgentSystem;
 
-using ICL.Core.Environment;
+using ICL.Core.Environments;
 
 using Rhino;
 using Rhino.Geometry;
