+++ ICL.GH/ICL.GH.csproj
@@ -32,6 +32,10 @@
     </Reference>
   </ItemGroup>
   
+  <ItemGroup>
+    <Folder Include="GhComponents\GhcEnvironment\" />
+  </ItemGroup>
+  
   <PropertyGroup Condition="$(Configuration) == 'Debug' AND $([MSBuild]::IsOSPlatform(Windows))">
     <StartProgram>C:\Program Files\Rhino 7\System\Rhino.exe</StartProgram>
     <StartArguments></StartArguments>
